수식 split 하는 부분 EquationParser로 분리
